Fixed test
Before it was running the exec method which had a bug, so the test was being tested with buggy code (we thought we were running a user defined setting of exec="foo" through .load, but it was actually exec="node"). This change ensures we're only passing through exactly the amount the user would.