changed --delay option: it now works like a throttle instead of with setTimeout.
This changes the behaviour of the option --delay n

This means that as soon as one file changes, nodemon will restart the
script. The previous behavior was to wait for n seconds before
restarting the process. With throttle, the function gets called when the
first file changes. After that, all file changes in the interval [t,t+n]
are ignored