fix: ignore bare directories fully
In the previous version, although the restart would ignore, an ignored
directory, if "bare" (i.e. without globs) would still be "monitored" and
potentially take up time. This change ensures the directory is fully
ignored (as well as a small change for fully ignoring files using the
same method).

Now the config.monitor rules is the source of truth for ignoring.

Fixes #1336