fix: restart on change for non-default signals (#1409)
When nodemon kills the child app in preparation for a restart, the app may exit
with a non-zero status code, especially when using custom signals like SIGINT.
Previously, this would be detected as a clean exit rather than a restart, at
least on macOS.

There was an existing flag, `killedAfterChange` which causes the desired
behavior, but it was only set on Windows. It seems that setting it always leaked
between tests and caused problems, so we only set it if a restart is expected.

Fixes #1409