fix: add download and extraction of windows-kill.exe v1.1.4 (#1720)
This commit adds an install script that runs on 'npm install'. Its
purpose is to download, extract and verify a specific version of
the windows-kill.exe binary.

I decided not add the - windows only - executable directly to this
project to keep the dependencies clear. Furthermore I did not use
the node-window-kill package, because it comes with node-gyp which
requires python and vs2017 build tools on the target machine.