fix: change child process termination for windows contexts (#1720)
This commit replaces the taskkill invocation with windows-kill.exe.

We are now sending a SIGINT POSIX signal to the node process in a
Windows context.

Known issues: The windows-kill.exe has to be triggered in a detached
process context, which currently flashes a terminal window on each
reload. Furthermore the 'restart' event is triggered twice on the bus.