fix: replace provisioning script with the raw windows-kill.exe (#1720)
This commit replaces the provisioning script, which downloaded
a specific windows-kill.exe binary, with the binary itself.
We are using the x64 version of the windows-kill.exe from:
https://github.com/ElyDotDev/windows-kill/releases/tag/1.1.4

We are now embedding the binary with the project directly to
decrease the complexity and avoid breaking the build by external
changes. Actually the maintainer of windows-kill changed his
GitHub username recently from "alirdn" to "ElyDotDev" which
broke the provisioning script!